+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.1
+
+- Fixed message id title.
+- Update template plugin to the last templates version.
+
 # 0.1.0
 
 - Added resolver.